fix factorial function for Unicode case correct as part of sourceforge rcracki_mt bug #2926357 move "Failed case correction" message into RainbowCrack.cpp and out of lm2ntlm.cpp

+++ client_applications/rcracki_mt/lm2ntlm.h
@@ -63,7 +63,7 @@ private:
 	bool checkPermutations(int length, unsigned char* pTempMute, int* jAtPos, int* sizeAtPos, unsigned char* pLMPassword, std::string& sNTLMPassword);
 
 	int calculateTotalCombinations(int length, int setSize);
-	int factorial (int num);
+	uint64 factorial (int num);
 
 	bool parseHexPassword(std::string hexPassword, std::string& sPlain);
 	bool NormalizeHexString(std::string& sHash);
